patch: Change `master` branch to `main` in README
Changed references to `master` branch to `main` in alignment with GitHub's shift to a new default.

GitHub repos now default to `main` instead of `master`. When I created a new repo which had a default branch of `main` and I used the example from the README the workflow didn't trigger because it was pointing to `master`. This change amends the README to refer to the new default, and changes the example action to trigger on both `master` and `main` branches to be backwards compatible. 
 
See [this article]( https://www.zdnet.com/article/github-to-replace-master-with-main-starting-next-month/#:~:text=Starting%20next%20month,%20all%20new,them%20with%20more%20inclusive%20terms) for changes to GitHub default.